requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
execnet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.2
pytest-xdist==3.3.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.6
webdriver-manager==4.0.1
wsproto==1.2.0
